Create NumberOfIslands.cpp
Time complexity:
O(mxn)

The algorithm iterates through each cell of the grid once. For each cell, if it's an unvisited '1', we perform a DFS/BFS traversal. Critically, each cell is visited at most twice: once by the outer loops and at most once by the DFS/BFS traversal. The DFS/BFS only explores each land cell of an island once. Therefore, the total time spent is proportional to the number of cells in the grid.
m = number of rows
n = number of columns

Space complexity:
O(mxn)

In the worst-case scenario, the entire grid could be a single island. The recursion depth of the DFS call stack could be as large as the total number of cells in the grid, i.e., m⋅n. This space is used to store the call stack frames. For a BFS approach, the queue could hold up to m⋅n elements.