Create KthMissingPositiveNumber.java
Time complexity: O(logn), where n is the length of the input array. We are using binary search to find the index of the kth missing positive integer, which takes logarithmic time.
Space complexity: O(1). We are using constant extra space to store the left and right pointers in binary search.